Extract our Step system out of cloud debug for reuse
