+++ jetbrains-core/src/software/aws/toolkits/jetbrains/services/clouddebug/DebuggerSupport.kt
@@ -20,10 +20,10 @@ import com.intellij.openapi.extensions.ExtensionPointName
 import com.intellij.util.execution.ParametersListUtil
 import com.intellij.util.io.isDirectory
 import com.intellij.util.io.isFile
-import software.aws.toolkits.jetbrains.services.clouddebug.execution.Context
 import software.aws.toolkits.jetbrains.services.clouddebug.execution.steps.ResourceTransferStep
 import software.aws.toolkits.jetbrains.services.ecs.execution.ImmutableArtifactMapping
 import software.aws.toolkits.jetbrains.services.ecs.execution.ImmutableContainerOptions
+import software.aws.toolkits.jetbrains.utils.execution.steps.Context
 import software.aws.toolkits.resources.message
 import java.io.File
 import java.nio.file.Paths
