+++ jetbrains-core/src/software/aws/toolkits/jetbrains/services/clouddebug/actions/PseCliAction.kt
@@ -40,12 +40,12 @@ import software.aws.toolkits.jetbrains.services.clouddebug.CliOutputParser
 import software.aws.toolkits.jetbrains.services.clouddebug.CloudDebugExecutable
 import software.aws.toolkits.jetbrains.services.clouddebug.CloudDebugResolver
 import software.aws.toolkits.jetbrains.services.clouddebug.asLogEvent
-import software.aws.toolkits.jetbrains.services.clouddebug.execution.DefaultMessageEmitter
 import software.aws.toolkits.jetbrains.services.clouddebug.resources.CloudDebuggingResources
 import software.aws.toolkits.jetbrains.services.ecs.EcsClusterNode
 import software.aws.toolkits.jetbrains.services.ecs.EcsServiceNode
 import software.aws.toolkits.jetbrains.services.ecs.resources.EcsResources.describeService
 import software.aws.toolkits.jetbrains.services.ecs.resources.EcsResources.listServiceArns
+import software.aws.toolkits.jetbrains.utils.execution.steps.MessageEmitter
 import software.aws.toolkits.jetbrains.utils.notifyError
 import software.aws.toolkits.jetbrains.utils.notifyInfo
 import software.aws.toolkits.resources.message
@@ -76,7 +76,7 @@ abstract class PseCliAction(val project: Project, val actionName: String, privat
                         "",
                         System.currentTimeMillis()
                     )
-                    val messageEmitter = DefaultMessageEmitter.createRoot(buildViewManager, actionName)
+                    val messageEmitter = MessageEmitter.createRoot(buildViewManager, actionName)
                     buildViewManager.onEvent(actionName, StartBuildEventImpl(descriptor, ""))
 
                     val toolWindowManager = ToolWindowManager.getInstance(project)
