+++ jetbrains-core/src/software/aws/toolkits/jetbrains/services/clouddebug/execution/CloudDebugRunState.kt
@@ -25,6 +25,8 @@ import software.aws.toolkits.jetbrains.services.clouddebug.execution.steps.RootS
 import software.aws.toolkits.jetbrains.services.clouddebug.execution.steps.SetUpPortForwarding
 import software.aws.toolkits.jetbrains.services.clouddebug.execution.steps.StopApplications
 import software.aws.toolkits.jetbrains.services.ecs.execution.EcsServiceCloudDebuggingRunSettings
+import software.aws.toolkits.jetbrains.utils.execution.steps.Context
+import software.aws.toolkits.jetbrains.utils.execution.steps.MessageEmitter
 import software.aws.toolkits.resources.message
 import software.aws.toolkits.telemetry.ClouddebugTelemetry
 import software.aws.toolkits.telemetry.Result
@@ -58,7 +60,7 @@ class CloudDebugRunState(
         val processHandler = CloudDebugProcessHandler(context)
 
         ApplicationManager.getApplication().executeOnPooledThread {
-            val messageEmitter = DefaultMessageEmitter.createRoot(buildView, runConfigId())
+            val messageEmitter = MessageEmitter.createRoot(buildView, runConfigId())
             var result = Result.Succeeded
             try {
                 startRunConfiguration(descriptor, buildView)
