+++ jetbrains-core/src/software/aws/toolkits/jetbrains/services/clouddebug/execution/steps/AttachDebuggers.kt
@@ -15,12 +15,12 @@ import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.application.runInEdt
 import com.intellij.openapi.util.Key
 import software.aws.toolkits.jetbrains.services.clouddebug.DebuggerSupport
-import software.aws.toolkits.jetbrains.services.clouddebug.execution.Context
-import software.aws.toolkits.jetbrains.services.clouddebug.execution.MessageEmitter
-import software.aws.toolkits.jetbrains.services.clouddebug.execution.ParallelStep
-import software.aws.toolkits.jetbrains.services.clouddebug.execution.Step
 import software.aws.toolkits.jetbrains.services.ecs.execution.EcsServiceCloudDebuggingRunSettings
 import software.aws.toolkits.jetbrains.services.ecs.execution.ImmutableContainerOptions
+import software.aws.toolkits.jetbrains.utils.execution.steps.Context
+import software.aws.toolkits.jetbrains.utils.execution.steps.MessageEmitter
+import software.aws.toolkits.jetbrains.utils.execution.steps.ParallelStep
+import software.aws.toolkits.jetbrains.utils.execution.steps.Step
 import software.aws.toolkits.resources.message
 import software.aws.toolkits.telemetry.CloudDebugPlatform
 import software.aws.toolkits.telemetry.ClouddebugTelemetry
