+++ jetbrains-core/src/software/aws/toolkits/jetbrains/services/clouddebug/execution/steps/StartApplications.kt
@@ -6,11 +6,11 @@ package software.aws.toolkits.jetbrains.services.clouddebug.execution.steps
 import com.intellij.execution.configurations.GeneralCommandLine
 import com.intellij.util.execution.ParametersListUtil
 import software.aws.toolkits.jetbrains.core.credentials.toEnvironmentVariables
-import software.aws.toolkits.jetbrains.services.clouddebug.execution.CliBasedStep
-import software.aws.toolkits.jetbrains.services.clouddebug.execution.Context
-import software.aws.toolkits.jetbrains.services.clouddebug.execution.ParallelStep
-import software.aws.toolkits.jetbrains.services.clouddebug.execution.Step
+import software.aws.toolkits.jetbrains.services.clouddebug.execution.CloudDebugCliStep
 import software.aws.toolkits.jetbrains.services.ecs.execution.EcsServiceCloudDebuggingRunSettings
+import software.aws.toolkits.jetbrains.utils.execution.steps.Context
+import software.aws.toolkits.jetbrains.utils.execution.steps.ParallelStep
+import software.aws.toolkits.jetbrains.utils.execution.steps.Step
 import software.aws.toolkits.resources.message
 import software.aws.toolkits.telemetry.ClouddebugTelemetry
 import software.aws.toolkits.telemetry.Result
@@ -30,23 +30,21 @@ class StartApplication(
     private val settings: EcsServiceCloudDebuggingRunSettings,
     private val containerName: String,
     private val startCommand: String
-) : CliBasedStep() {
+) : CloudDebugCliStep() {
     override val stepName = message("cloud_debug.step.start_application.resource", containerName)
 
-    override fun constructCommandLine(context: Context, commandLine: GeneralCommandLine) {
-        commandLine
-            .withParameters("--verbose")
-            .withParameters("--json")
-            .withParameters("start")
-            .withParameters("--target")
-            .withParameters(ResourceInstrumenter.getTargetForContainer(context, containerName))
-            /* TODO remove this when the cli conforms to the contract */
-            .withParameters("--selector")
-            .withParameters(containerName)
-            .withParameters(ParametersListUtil.parse(startCommand, false, false))
-            .withEnvironment(settings.region.toEnvironmentVariables())
-            .withEnvironment(settings.credentialProvider.resolveCredentials().toEnvironmentVariables())
-    }
+    override fun constructCommandLine(context: Context): GeneralCommandLine = getCli(context)
+        .withParameters("--verbose")
+        .withParameters("--json")
+        .withParameters("start")
+        .withParameters("--target")
+        .withParameters(ResourceInstrumenter.getTargetForContainer(context, containerName))
+        /* TODO remove this when the cli conforms to the contract */
+        .withParameters("--selector")
+        .withParameters(containerName)
+        .withParameters(ParametersListUtil.parse(startCommand, false, false))
+        .withEnvironment(settings.region.toEnvironmentVariables())
+        .withEnvironment(settings.credentialProvider.resolveCredentials().toEnvironmentVariables())
 
     override fun recordTelemetry(context: Context, startTime: Instant, result: Result) {
         ClouddebugTelemetry.startApplication(
