+++ jetbrains-core/src/software/aws/toolkits/jetbrains/services/clouddebug/execution/steps/StopApplications.kt
@@ -5,12 +5,12 @@ package software.aws.toolkits.jetbrains.services.clouddebug.execution.steps
 
 import com.intellij.execution.configurations.GeneralCommandLine
 import software.aws.toolkits.jetbrains.core.credentials.toEnvironmentVariables
-import software.aws.toolkits.jetbrains.services.clouddebug.execution.CliBasedStep
-import software.aws.toolkits.jetbrains.services.clouddebug.execution.Context
-import software.aws.toolkits.jetbrains.services.clouddebug.execution.MessageEmitter
-import software.aws.toolkits.jetbrains.services.clouddebug.execution.ParallelStep
-import software.aws.toolkits.jetbrains.services.clouddebug.execution.Step
+import software.aws.toolkits.jetbrains.services.clouddebug.execution.CloudDebugCliStep
 import software.aws.toolkits.jetbrains.services.ecs.execution.EcsServiceCloudDebuggingRunSettings
+import software.aws.toolkits.jetbrains.utils.execution.steps.Context
+import software.aws.toolkits.jetbrains.utils.execution.steps.MessageEmitter
+import software.aws.toolkits.jetbrains.utils.execution.steps.ParallelStep
+import software.aws.toolkits.jetbrains.utils.execution.steps.Step
 import software.aws.toolkits.resources.message
 import software.aws.toolkits.telemetry.ClouddebugTelemetry
 import software.aws.toolkits.telemetry.Result
@@ -37,26 +37,24 @@ class StopApplication(
     private val settings: EcsServiceCloudDebuggingRunSettings,
     private val containerName: String,
     private val isCleanup: Boolean
-) : CliBasedStep() {
+) : CloudDebugCliStep() {
     override val stepName = if (isCleanup) {
         message("cloud_debug.step.stop_application.cleanup.resource", containerName)
     } else {
         message("cloud_debug.step.stop_application.pre_start.resource", containerName)
     }
 
-    override fun constructCommandLine(context: Context, commandLine: GeneralCommandLine) {
-        commandLine
-            .withParameters("--verbose")
-            .withParameters("--json")
-            .withParameters("stop")
-            .withParameters("--target")
-            .withParameters(ResourceInstrumenter.getTargetForContainer(context, containerName))
-            /* TODO remove this when the cli conforms to the contract */
-            .withParameters("--selector")
-            .withParameters(containerName)
-            .withEnvironment(settings.region.toEnvironmentVariables())
-            .withEnvironment(settings.credentialProvider.resolveCredentials().toEnvironmentVariables())
-    }
+    override fun constructCommandLine(context: Context): GeneralCommandLine = getCli(context)
+        .withParameters("--verbose")
+        .withParameters("--json")
+        .withParameters("stop")
+        .withParameters("--target")
+        .withParameters(ResourceInstrumenter.getTargetForContainer(context, containerName))
+        /* TODO remove this when the cli conforms to the contract */
+        .withParameters("--selector")
+        .withParameters(containerName)
+        .withEnvironment(settings.region.toEnvironmentVariables())
+        .withEnvironment(settings.credentialProvider.resolveCredentials().toEnvironmentVariables())
 
     override fun recordTelemetry(context: Context, startTime: Instant, result: Result) {
         ClouddebugTelemetry.stopApplication(
