+++ jetbrains-core/src/software/aws/toolkits/jetbrains/utils/execution/steps/MessageEmitter.kt
@@ -1,7 +1,7 @@
-// Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package software.aws.toolkits.jetbrains.services.clouddebug.execution
+package software.aws.toolkits.jetbrains.utils.execution.steps
 
 import com.intellij.build.BuildProgressListener
 import com.intellij.build.events.impl.FailureResultImpl
@@ -10,33 +10,24 @@ import com.intellij.build.events.impl.OutputBuildEventImpl
 import com.intellij.build.events.impl.StartEventImpl
 import com.intellij.build.events.impl.SuccessResultImpl
 
-interface MessageEmitter {
-    fun startStep()
-    fun emitMessage(message: String, isError: Boolean)
-    fun finishSuccessfully()
-    fun finishExceptionally(e: Throwable)
-    fun createChild(stepName: String, hidden: Boolean = false): MessageEmitter
-}
-
-class DefaultMessageEmitter private constructor(
+class MessageEmitter private constructor(
     private val buildListener: BuildProgressListener,
     private val rootObject: Any,
     private val parentId: String,
     private val stepName: String,
     private val hidden: Boolean,
     private val parent: MessageEmitter?
-) : MessageEmitter {
-
-    override fun createChild(stepName: String, hidden: Boolean): DefaultMessageEmitter {
+) {
+    fun createChild(stepName: String, hidden: Boolean = false): MessageEmitter {
         val (childParent, childStepName) = if (hidden) {
             parentId to this.stepName
         } else {
             this.stepName to stepName
         }
-        return DefaultMessageEmitter(buildListener, rootObject, childParent, childStepName, hidden, this)
+        return MessageEmitter(buildListener, rootObject, childParent, childStepName, hidden, this)
     }
 
-    override fun startStep() {
+    fun startStep() {
         if (hidden) return
         buildListener.onEvent(
             rootObject,
@@ -49,7 +40,7 @@ class DefaultMessageEmitter private constructor(
         )
     }
 
-    override fun finishSuccessfully() {
+    fun finishSuccessfully() {
         if (hidden) return
         buildListener.onEvent(
             rootObject,
@@ -63,7 +54,7 @@ class DefaultMessageEmitter private constructor(
         )
     }
 
-    override fun finishExceptionally(e: Throwable) {
+    fun finishExceptionally(e: Throwable) {
         emitMessage("$stepName finished exceptionally: $e", true)
         if (hidden) return
         buildListener.onEvent(
@@ -78,7 +69,7 @@ class DefaultMessageEmitter private constructor(
         )
     }
 
-    override fun emitMessage(message: String, isError: Boolean) {
+    fun emitMessage(message: String, isError: Boolean) {
         parent?.emitMessage(message, isError)
         if (hidden) return
         buildListener.onEvent(
@@ -93,6 +84,6 @@ class DefaultMessageEmitter private constructor(
 
     companion object {
         fun createRoot(buildListener: BuildProgressListener, rootStepName: String, hidden: Boolean = false): MessageEmitter =
-            DefaultMessageEmitter(buildListener, rootStepName, rootStepName, rootStepName, hidden, null)
+            MessageEmitter(buildListener, rootStepName, rootStepName, rootStepName, hidden, null)
     }
 }
