+++ resources/resources/software/aws/toolkits/resources/localized_messages.properties
@@ -401,6 +401,9 @@ feedback.validation.no_sentiment=Please select how you're feeling.
 general.close_button=Close
 general.configure_button=Configure
 general.create_button=Create
+general.execution.cancelled=canceled
+general.execution.failed=failed
+general.execution.success=completed successfully
 general.file_not_found=File not found: "{0}"
 general.in_progress_button=In progress
 general.message=Message
